+++ CHANGES.md
@@ -13,6 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 - Replace usage of `wkb` crate with `geozero` because of licensing.
 - Stop supporting Python 3.6 (was EOL on 23 Dec 2021). Minimum supported python version now is Python 3.7.
+- Omit empty geometries when converting vector data.
 
 ## h3ronpy [0.13.1] - 2021-12-13